Ensure that puppetlabs_yum doesn't do crazy things on top of PE.
Prior to this commit, runnign this module on a system running Puppet Enterprise
would just install the FOSS repos. This can lead to unpredictable behavior for
PE users, and generally non-goodness.

This commit creates a custom fact to determine if pe-puppet-enterprise-release
is installed. If it is, we just noop the module. This allows for all systems to
have this modules setup and classified while still retraining relatively sane
and expected behavior.