If pper_installed is unef, fall through to FOSS error message
Prior to this commit, if you were on a non-yum using system you would
get a mesage saying you were not configuring puppetlabs_yum because you
had Puppet Enterprise installed. This simply wasn't true. Logic now
fixed to ensure if you're on a non-yum system it errors out with a
correct message.

This resolves and item brough up in issue #5.