Stop hard-coding (even with variables) Yum configuration
Prior to this commit, we were figuring out the $os_maj_verion and building that
in as the url parameter in the baseurl for the os version. This was fine until
you wanted to do a rolling upgrade all in one bang (using something like
fedup). This fix should reset the url parameter to be yum's $releasever rather
than a static string that puppet lays down.